changes done, moved to utils and used tailwind
