+++ src/components/Patient/SampleTestCard.tsx
@@ -12,8 +12,7 @@ import { NonReadOnlyUsers } from "@/Utils/AuthorizeFor";
 import * as Notification from "@/Utils/Notifications";
 import routes from "@/Utils/request/api";
 import request from "@/Utils/request/request";
-import { camelCase, startCase } from "@/Utils/stringUtils";
-import { formatDateTime } from "@/Utils/utils";
+import { camelCase, formatDateTime } from "@/Utils/utils";
 
 interface SampleDetailsProps {
   facilityId: number;
@@ -101,7 +100,9 @@ export const SampleTestCard = (props: SampleDetailsProps) => {
               Status{" "}
             </div>
             <div className="mt-1 overflow-x-scroll whitespace-normal break-words text-sm font-medium leading-5">
-              {startCase(camelCase(itemData.status || ""))}
+              <span className="capitalize">
+                {camelCase(itemData.status || "")}
+              </span>
             </div>
           </div>
         </div>
@@ -133,7 +134,9 @@ export const SampleTestCard = (props: SampleDetailsProps) => {
               Result{" "}
             </div>
             <div className="mt-1 overflow-x-scroll whitespace-normal break-words text-sm font-medium leading-5">
-              {startCase(camelCase(itemData.result || ""))}
+              <span className="capitalize">
+                {camelCase(itemData.result || "")}
+              </span>
             </div>
           </div>
         </div>
